Update Wolf3d and DoomEngines
