+++ ports/doomengines/doomengines/configs/gzdoom/gzdoom.ini
@@ -6,7 +6,7 @@ Path=.
 Path=$DOOMWADDIR
 Path=/storage/.config/gzdoom
 Path=/storage/roms/doom
-Path=../iwads
+Path=./iwads
 
 # These are the directories to search for wads added with the -file
 # command line parameter, if they cannot be found with the path
@@ -15,15 +15,15 @@ Path=../iwads
 Path=/storage/.config/gzdoom
 Path=/storage/roms/doom
 Path=$DOOMWADDIR
-Path=../iwads
+Path=./iwads
 
 # These are the directories to search for soundfonts that let listed in the menu.
 # Layout is the same as for IWADSearch.Directories
 [SoundfontSearch.Directories]
 Path=/storage/.config/gzdoom/soundfonts
 Path=/storage/.config/gzdoom/fm_banks
-Path=.engines/soundfonts
-Path=.engines/fm_banks
+Path=./engines/shared/soundfonts
+Path=./engines/shared/fm_banks
 
 # Files to automatically execute when running the corresponding game.
 # Each file should be on its own line, preceded by Path=
@@ -347,7 +347,7 @@ r_skipmats=false
 r_sprite_distance_cull=0
 r_spriteadjust=2
 r_ticstability=true
-save_dir=
+save_dir=./configs/gzdoom/saves
 save_formatted=false
 saveloadconfirmation=true
 savestatistics=0
