+++ ports/doomengines/doomengines/configs/gzdoom/heretic.ini
@@ -354,7 +354,7 @@ r_skipmats=false
 r_sprite_distance_cull=0
 r_spriteadjust=2
 r_ticstability=true
-save_dir=
+save_dir=./configs/gzdoom/saves
 save_formatted=false
 saveloadconfirmation=true
 savestatistics=0
