+++ ports/wolf3d/gameinfo.xml
@@ -1,5 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
-<gameList>
 	<game>
 		<path>./Wolfenstein 3D.sh</path>
 		<name>Wolfenstein 3D</name>
@@ -10,4 +8,3 @@
 		<publisher>Various</publisher>
 		<genre>Shooter</genre>
 	</game>
-</gameList>
